services/ai-service/src/main.py
from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, List
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="AI Service", version="1.0.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Health check
@app.get("/health")
async def health():
    return {"status": "healthy", "service": "ai-service"}

# Request models
class ChatMessage(BaseModel):
    message: str
    language: str = "en"
    user_id: Optional[str] = None

class RecommendationRequest(BaseModel):
    user_id: str
    context: Optional[dict] = None

class YieldPredictionRequest(BaseModel):
    crop_type: str
    farm_size: float
    location: dict
    historical_yield: Optional[List[float]] = None

class PricePredictionRequest(BaseModel):
    crop_type: str
    location: dict
    quantity: float

# Chat endpoint
@app.post("/api/chat")
async def chat(request: ChatMessage):
    """
    AI-powered chatbot for farming questions
    """
    try:
        # TODO: Implement multilingual BERT-based chatbot
        # This is a placeholder response
        response = {
            "message": f"AI Assistant response to: {request.message}",
            "language": request.language,
            "confidence": 0.85
        }
        return response
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Image analysis endpoint
@app.post("/api/image/analyze")
async def analyze_image(file: UploadFile = File(...)):
    """
    Analyze uploaded image for pest/disease detection
    """
    try:
        # TODO: Implement computer vision model
        # Save uploaded file temporarily
        contents = await file.read()
        
        # Placeholder response
        response = {
            "detection_type": "pest",
            "confidence": 0.92,
            "detected_items": [
                {
                    "name": "Aphids",
                    "confidence": 0.92,
                    "severity": "moderate",
                    "recommendations": [
                        "Use neem oil spray",
                        "Introduce ladybugs",
                        "Remove affected leaves"
                    ]
                }
            ],
            "plant_health": "moderate"
        }
        return response
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Recommendations endpoint
@app.post("/api/recommendations")
async def get_recommendations(request: RecommendationRequest):
    """
    Get personalized recommendations for missions, crops, etc.
    """
    try:
        # TODO: Implement recommendation engine
        recommendations = {
            "missions": [
                {
                    "id": "mission_1",
                    "title": "Compost Setup",
                    "reason": "Based on your farm type",
                    "priority": "high"
                }
            ],
            "crops": [
                {
                    "crop": "Rice",
                    "reason": "Suitable for your region",
                    "season": "Kharif"
                }
            ]
        }
        return recommendations
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Yield prediction endpoint
@app.post("/api/predict/yield")
async def predict_yield(request: YieldPredictionRequest):
    """
    Predict crop yield based on various factors
    """
    try:
        # TODO: Implement XGBoost/Prophet model
        prediction = {
            "predicted_yield": 2500.5,  # kg per hectare
            "confidence_interval": {
                "lower": 2300.0,
                "upper": 2700.0
            },
            "factors": {
                "weather_impact": "positive",
                "soil_quality": "good",
                "pest_risk": "low"
            }
        }
        return prediction
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Price prediction endpoint
@app.post("/api/predict/price")
async def predict_price(request: PricePredictionRequest):
    """
    Predict market price for crops
    """
    try:
        # TODO: Implement LSTM-based price forecasting
        prediction = {
            "predicted_price": 45.50,  # per kg
            "currency": "INR",
            "confidence": 0.78,
            "trend": "increasing",
            "best_sell_time": "2024-03-15"
        }
        return prediction
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=3003)


