services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
torch==2.1.0
transformers==4.35.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.3
redis==5.0.1
httpx==0.25.1
python-dotenv==1.0.0
mlflow==2.8.1
pydantic-settings==2.1.0


